Move Chart to separate folder and use it in bothh components
