+++ salty-trolls-react/src/components/Hacker/Hacker.scss
@@ -3,7 +3,7 @@
   width: 100%;
 }
 .hacker a {
-  margin: 10px 0;
+  margin: 20px 0;
   width: 100%;
   background-color: white;
   color: black;
