+++ salty-trolls-react/src/components/HackerProfile/HackerProfile.scss
@@ -14,7 +14,7 @@
   margin: 20px;
   padding: 30px 0 0 70px;
   display: flex;
-  height: 230px;
+  height: 260px;
   background: #ffff;
   color: black;
   align-items: center;
@@ -23,6 +23,7 @@
 
 .details-left-box {
   text-align: center;
+  margin-top: -30px;
 }
 .icon {
   width: 22%;
